Split delivery recovery cursor planning into stable seam
Extract delivery recovery cursor planning/state helpers from
adapters/discord/service.py into a new delivery_recovery.py module
with a narrow public API (plan_delivery_recovery_cursor,
is_recovery_backoff_active). Mechanical move only; no behavior changes.

Also fix basetemp cleanup in conftest.py to handle xdist worker
teardown races, and update test_interaction_runtime_chaos.py import
for the moved constant.
